[Obs AI Assistant] Support querying `semantic_text` fields in search connectors (#200184)
This adds support for querying search connectors that have
`semantic_text` fields.

---------
