+++ x-pack/plugins/observability_solution/observability_ai_assistant/tsconfig.json
@@ -46,6 +46,7 @@
     "@kbn/management-settings-ids",
     "@kbn/ai-assistant-common",
     "@kbn/inference-common",
+    "@kbn/core-lifecycle-server",
   ],
   "exclude": ["target/**/*"]
 }
